Remove isRootRelation knowledge from consumers
This commit changes the consumers contract.
A consumer now plans as distributed as possible and no longer cares if
it has a parent relation or not.

Any information that is necessary to merge this distributed result has
to be provided in a ResultDescription which can be picked up by a parent
to create a handler-merge.

This changes some execution plans slightly. For example, an insert with
a distributed group by in a subquery will now execute the inserts from
the reducer nodes instead of from the handler node.